pyproject.toml
[project]
name = "finbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "google-search-results>=2.4.2",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.2",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.42.0",
    "wikipedia>=1.4.0",
    "yfinance>=1.2.0",
]
